Add green/blue text bubbles
