+++ src/main/java/duke/DialogBox.java
@@ -1,17 +1,14 @@
 package duke;
-import java.io.IOException;
-import java.util.Collections;
 
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
+import java.io.IOException;
+
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.HBox;
+
 
 /**
  * An example of a custom control using FXML.
@@ -37,38 +34,9 @@ public class DialogBox extends HBox {
         dialog.setText(text);
         displayPicture.setImage(img);
     }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        Collections.reverse(tmp);
-        getChildren().setAll(tmp);
-        setAlignment(Pos.TOP_LEFT);
-    }
-
     public static DialogBox getUserDialog(String text, Image img) {
-        return new DialogBox(text, img);
-    }
-
-    public static DialogBox getDukeDialog(String text, Image img) {
-        var db = new DialogBox(text, img);
-        db.flip();
+        DialogBox db = new DialogBox(text, img);
         return db;
-    }
-
-    /**
-     * Greeting dialog box.
-     *
-     * @param iv the image to be displayed in the greeting box.
-     * @return the dialog box.
-     */
-    public static DialogBox greetingDialogBox(Image iv) {
-        String greeting = "Welcome! I am Tom!";
-        var duke = new DialogBox(greeting, iv);
-        duke.flip();
-        return duke;
     }
 }
