+++ src/main/resources/view/DukeReplyDialogBox.fxml
@@ -7,7 +7,7 @@
 
 <fx:root alignment="TOP_RIGHT" maxHeight="-Infinity" maxWidth="1.7976931348623157E308" prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <Label fx:id="dialog" maxHeight="-Infinity" minHeight="-Infinity" style="-fx-label-padding: 10; -fx-background-radius: 15; -fx-background-color: #80b1ff;" text="Label" wrapText="true" />
+        <Label fx:id="dialog" maxHeight="-Infinity" minHeight="-Infinity" style="-fx-label-padding: 10; -fx-background-radius: 15; -fx-background-color: #90ee90;" text="Label" wrapText="true" />
         <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true" />
     </children>
     <padding>
